refactor: Move server to internal package
