+++ internal/server/handler/serialize.go
