+++ internal/server/handler/v1/errors.go
