+++ internal/server/handler/v1/handler_v1.go
@@ -8,8 +8,8 @@ import (
 
 	v1 "github.com/glass-cms/glasscms/api/v1"
 	"github.com/glass-cms/glasscms/internal/item"
+	"github.com/glass-cms/glasscms/internal/server/handler"
 	"github.com/glass-cms/glasscms/pkg/resource"
-	"github.com/glass-cms/glasscms/server/handler"
 )
 
 type APIHandler struct {
